+++ src/core/transaction.ts
@@ -1,5 +1,6 @@
 import {globalState} from "./globalstate";
 import {runReactions} from "./reaction";
+import {startBatch, endBatch} from "./observable";
 import {isSpyEnabled, spyReportStart, spyReportEnd} from "./spy";
 
 /**
@@ -16,6 +17,7 @@ export function transaction<T>(action: () => T, thisArg = undefined, report = tr
 }
 
 export function transactionStart<T>(name: string, thisArg = undefined, report = true) {
+	startBatch();
 	globalState.inTransaction += 1;
 	if (report && isSpyEnabled()) {
 		spyReportStart({
@@ -32,4 +34,5 @@ export function transactionEnd<T>(report = true) {
 	}
 	if (report && isSpyEnabled())
 		spyReportEnd();
+	endBatch();
 }