+++ src/mobx.ts
@@ -19,9 +19,9 @@
 import {registerGlobals} from "./core/globalstate";
 registerGlobals();
 
-export { IAtom, Atom                                          } from "./core/atom";
+export { IAtom, Atom, BaseAtom                                } from "./core/atom";
 export { IObservable, IDepTreeNode, DerivationsSets           } from "./core/observable";
-export { Reaction                                             } from "./core/reaction";
+export { Reaction, IReactionPublic                            } from "./core/reaction";
 export { IDerivation, untracked                               } from "./core/derivation";
 export { action, useStrict, isAction, runInAction             } from "./core/action";
 export { spy                                                  } from "./core/spy";