+++ src/types/modifiers.ts
@@ -120,7 +120,7 @@ export function makeChildObservable(value, parentMode: ValueMode, name?: string)
 	}
 
 	if (Array.isArray(value))
-		return createObservableArray(<[]> value, childMode, name);
+		return createObservableArray(value as any[], childMode, name);
 	if (isPlainObject(value) && Object.isExtensible(value))
 		return extendObservableHelper(value, value, childMode, name);
 	return value;